test: move the can_run tests next to ref.py
They cover builtin/ref.py, so test_api.py was the wrong mirror - that
file belongs to via_global_ref/api.py and nothing else tested ref.py.

Configure the mocked stat after building the autospec: create_autospec
applies dotted kwargs only from 3.12 on, so on 3.9-3.11 st_mode stayed a
MagicMock and every can_run assertion saw True.
